+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Fixed
+- Fix selection process for file matrix 
+- Allow multi-packages to be CommonJS
+
+### Removed
+- Stop testing `package-lock.json` since it is a generated file